+++ pathmanager.py
@@ -10,7 +10,7 @@ Usage:
     python pathmanager.py --gui     # Run in GUI mode
     python pathmanager.py --help    # Show help message
 
-Version: v0.2.0c 2026-01-08 1645 CST
+Version: v0.2.0d 2026-01-09 1400 CST
 Author: Jim U'Ren
 License: Personal and educational use
 """
@@ -39,7 +39,7 @@ def main():
     parser.add_argument(
         "--version",
         action="version",
-        version="PathManager v0.2.0c 2026-01-08 1645 CST"
+        version="PathManager v0.2.0d 2026-01-09 1400 CST"
     )
 
     args = parser.parse_args()